changed tests and models to fit admin panel

+++ webapp/authapp/models.py
@@ -6,43 +6,8 @@ from django.core import validators
 from django.core.mail import send_mail
 from django.contrib.auth.models import BaseUserManager, AbstractUser
 import uuid
-from webapp.authapp.models import AppUser
 
 # Create your models here.
-
-
-class Telephone(models.Model):
-    user = models.OneToOneField(
-        AppUser,
-        on_delete=models.CASCADE,
-        primary_key=True
-    )
-    telephone_prefix = models.IntegerField(
-        "telephone number prefix",
-        default=48,
-        # https://en.wikipedia.org/wiki/List_of_country_calling_codes
-        validators=[
-            validators.MinValueValidator(1, "Minimum value must be 1"),
-            validators.MaxValueValidator(999, "Maximum value must be 999")
-        ],
-        help_text=''
-    )
-    telephone_number = models.CharField(
-        "telephone number",
-        max_length=9,
-        validators=[
-            validators.MinLengthValidator(9, "Invalid telephone value")
-        ],
-        help_text=''
-    )
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=['telephone_prefix', 'telephone_number'],
-                name='telephone_number_with_prefix_unique'
-            )
-        ]
 
 
 class AppUserManager(BaseUserManager):
@@ -111,6 +76,33 @@ class AppUser(AbstractUser):
         primary_key=True,
         max_length=64
     )
+    telephone_prefix = models.IntegerField(
+        "telephone number prefix",
+        default=48,
+        # https://en.wikipedia.org/wiki/List_of_country_calling_codes
+        validators=[
+            validators.MinValueValidator(1, "Minimum value must be 1"),
+            validators.MaxValueValidator(999, "Maximum value must be 999")
+        ],
+        help_text=''
+    )
+    telephone_number = models.CharField(
+        "telephone number",
+        max_length=9,
+        null=True,
+        validators=[
+            validators.MinLengthValidator(9, "Invalid telephone value")
+        ],
+        help_text=''
+    )
+
+    class Meta:
+        constraints = [
+            models.UniqueConstraint(
+                fields=['telephone_prefix', 'telephone_number'],
+                name='telephone_number_with_prefix_unique'
+            )
+        ]
 
     USERNAME_FIELD = 'email'
     REQUIRED_FIELDS = []
@@ -136,30 +128,102 @@ class AppUser(AbstractUser):
         return send_mail(subject, message, from_email, [self.email], **kwargs)
 
     @staticmethod
-    def create_admin(email: str, password: str):
-        '''
-        Create admin account.
-        '''
-        pwd_hashed = make_password(password)
+    def create_user(
+        email: str,
+        telephone_number: str,
+        telephone_prefix: Optional[int] = None,
+        delete_if_mail_not_exists: bool = True
+    ) -> 'AppUser':
+        """
+        Create normal user account and send e-mail with password.
 
-        AppUser.objects.create_superuser(
+        Args:
+            email: admin email address
+            telephone_prefix: user telephone prefix
+            telephone_number: user telephone number
+            delete_if_mail_not_exists: if mail not exists, then remove account
+
+        Returns:
+            AppUser: admin account object
+
+        Raises:
+            AppUser.InvalidEmail: E-mail with password cannot be send
+        """
+        return AppUser._create_account(
             email=email,
-            password=pwd_hashed
+            is_admin=True,
+            telephone_number=telephone_number,
+            telephone_prefix=telephone_prefix,
+            delete_if_mail_not_exists=delete_if_mail_not_exists
         )
 
     @staticmethod
-    def create_user(email: str, telephone: Optional[Telephone] = None):
-        '''
-        Create hashed unique password and send email to the user with password.
-        '''
-        password = uuid.uuid4().hex
-        pwd_hashed = make_password(password)
+    def create_admin(
+        email: str,
+        delete_if_mail_not_exists: bool = True
+    ) -> 'AppUser':
+        """
+        Create admin user account and send e-mail with password.
 
-        user = AppUser.objects.create_user(
+        Args:
+            email: admin email address
+            delete_if_mail_not_exists: if mail not exists, then remove account
+
+        Returns:
+            AppUser: admin account object
+
+        Raises:
+            AppUser.InvalidEmail: E-mail with password cannot be send
+        """
+        return AppUser._create_account(
             email=email,
-            password=pwd_hashed,
-            telephone=telephone
+            is_admin=True,
+            delete_if_mail_not_exists=delete_if_mail_not_exists
         )
+
+    @staticmethod
+    def _create_account(
+        email: str,
+        is_admin: bool,
+        telephone_prefix: Optional[int] = None,
+        telephone_number: Optional[str] = None,
+        delete_if_mail_not_exists: bool = True
+    ) -> 'AppUser':
+        """
+        Create user account with unique hashed password, send e-mail with 
+        this password.
+
+        Args:
+            email: user email address (and ID) to send password
+            is_admin: sets user as superuser
+            telephone_prefix: user telephone prefix
+            telephone_number: user telephone number
+            delete_if_mail_not_exists: if mail not exists, then remove account
+
+        Returns:
+            AppUser: user account object
+
+        Raises:
+            AppUser.InvalidEmail: E-mail with password cannot be send
+        """
+        password = uuid.uuid4().hex
+
+        to_call = AppUser.objects.create_superuser \
+            if is_admin \
+            else AppUser.objects.create_user
+
+        # set prefix to field default value
+        if not telephone_prefix:
+            field = AppUser._meta.get_field('telephone_prefix')
+            telephone_prefix = field.default
+
+        user = to_call(
+            email,
+            password,
+            telephone_prefix=telephone_prefix,
+            telephone_number=telephone_number
+        )
+
         is_email_sent = user.email_user_with_status(
             'Account created',
             f'Password: {password}',
@@ -167,6 +231,9 @@ class AppUser(AbstractUser):
             fail_silently=False
         )
         if not is_email_sent:
-            user.delete()
+            if delete_if_mail_not_exists:
+                user.delete()
             msg = 'E-mail probably not exists - no password sent'
             raise AppUser.InvalidEmail(msg)
+
+        return user
